+++ src/pylibreconcile/__init__.py
@@ -1,4 +1,5 @@
-from .core import DesiredState
+from .change import Change, ChangeType
+from .desired_state import DesiredState
 from .known_state import (
     AWSS3KnownStateHandler,
     AzureStorageKnownStateHandler,
@@ -9,15 +10,25 @@ from .known_state import (
     SQLiteKnownStateHandler,
 )
 from .reconciler import Reconciler
+from .wiring import (
+    WiringContainer,
+    register_observed_state_handler,
+    register_resource_manager,
+)
 
 __all__ = [
     "AWSS3KnownStateHandler",
     "AzureStorageKnownStateHandler",
     "BoltDBKnownStateHandler",
+    "Change",
+    "ChangeType",
     "DesiredState",
     "KnownStateHandler",
     "LocalJSONKnownStateHandler",
     "LocalYAMLKnownStateHandler",
     "Reconciler",
     "SQLiteKnownStateHandler",
+    "WiringContainer",
+    "register_observed_state_handler",
+    "register_resource_manager",
 ]