+++ src/pylibreconcile/observed_state/__init__.py
@@ -1,0 +1,5 @@
+from .protocol import ObservedStateHandler as ObservedStateHandler
+
+__all__ = [
+    "ObservedStateHandler",
+]