+++ src/pylibreconcile/resource_manager/__init__.py
@@ -1,0 +1,5 @@
+from .protocol import ResourceManager as ResourceManager
+
+__all__ = [
+    "ResourceManager",
+]